+++ src/api/resources/construe/errors/index.ts
@@ -1,6 +1,7 @@
 export * from "./BadRequestError.js";
 export * from "./ConflictError.js";
 export * from "./FailedDependencyError.js";
+export * from "./ForbiddenError.js";
 export * from "./InternalServerError.js";
 export * from "./NotFoundError.js";
 export * from "./NotImplementedError.js";